+++ contracts/MultiResource_EIP/library/MultiResourceLib.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.0;
 
 library MultiResourceLib {
 
-  function removeItemByValue(uint32[] storage array, uint32 value) internal {
-    uint32[] memory memArr = array; //Copy array to memory, check for gas savings here
+  function removeItemByValue(uint64[] storage array, uint64 value) internal {
+    uint64[] memory memArr = array; //Copy array to memory, check for gas savings here
     uint256 length = memArr.length; //gas savings
     for (uint i; i<length; i = u_inc(i)) {
       if (memArr[i] == value) {
@@ -16,7 +16,7 @@ library MultiResourceLib {
   }
 
   //For reasource storage array
-  function removeItemByIndex(uint32[] storage array, uint256 index) internal {
+  function removeItemByIndex(uint64[] storage array, uint256 index) internal {
     //Check to see if this is already gated by require in all calls
     require(index < array.length);
     array[index] = array[array.length-1];
@@ -24,7 +24,7 @@ library MultiResourceLib {
   }
 
   //For reasource storage array
-  function removeItemByIndex(uint64[] storage array, uint256 index) internal {
+  function removeItemByIndex(uint128[] storage array, uint256 index) internal {
     //Check to see if this is already gated by require in all calls
     require(index < array.length);
     array[index] = array[array.length-1];
